pyproject.toml
[project]
name = "joint-project"
version = "0.1.0"
description = ""
authors = [
    {name = "pcm422",email = "pcm0422@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
